Extract contribution editor controller
Moves the WebView-based contribution editor hosting logic out of `ContributionDocumentView` into a dedicated `ContributionEditorController`, leaving the view as a thin adapter. This separates document lifecycle concerns from WebView, RPC, save, reload, and bridge management so the editor host can be reused by other surfaces.
