+++ Source/Workspace/Celbridge.Documents/Views/ContributionDocumentView.xaml.cs
@@ -1,113 +1,21 @@
-using System.Text.Json;
-using Celbridge.Commands;
-using Celbridge.DataTransfer;
-using Celbridge.Dialog;
+using Celbridge.Documents;
 using Celbridge.Documents.ViewModels;
-using Celbridge.Explorer;
-using Celbridge.Host;
-using Celbridge.Logging;
 using Celbridge.Messaging;
 using Celbridge.Packages;
-using Celbridge.Server;
-using Celbridge.UserInterface;
-using Celbridge.WebHost;
-using Celbridge.WebHost.Services;
 using Celbridge.Workspace;
-using Microsoft.Extensions.Localization;
-using Microsoft.Web.WebView2.Core;
-using Windows.ApplicationModel.DataTransfer;
 
 namespace Celbridge.Documents.Views;
 
 /// <summary>
-/// Document view for contribution-based editors, hosted via a WebView2.
+/// Document view for contribution-based editors, hosted via a WebView2. A thin adapter over
+/// ContributionEditorController: it satisfies the DocumentView contract (open, save timer, close, editor-state
+/// persistence) and forwards the WebView-hosting work to the controller.
 /// </summary>
-public sealed partial class ContributionDocumentView : DocumentView, IHostInput, IHostContext, IEditTarget
+public sealed partial class ContributionDocumentView : DocumentView
 {
-    private const int SaveRequestTimeoutSeconds = 30;
-    private const int ReloadStateWaitSeconds = 5;
-
-    // Editor-state capture/restore is best-effort (a user convenience, not data). Bound the wait so an
-    // editor that never responds to the host->editor RPC cannot stall document close, which would jam
-    // the serial command queue. Kept under the command watchdog's 5s threshold.
-    private const int EditorStateRequestTimeoutSeconds = 3;
-
-    private readonly ILogger<ContributionDocumentView> _logger;
-    private readonly ICommandService _commandService;
     private readonly IMessengerService _messengerService;
-    private readonly IStringLocalizer _stringLocalizer;
-    private readonly IDialogService _dialogService;
-    private readonly IUserInterfaceService _userInterfaceService;
-    private readonly IServiceProvider _serviceProvider;
-    private readonly IWebViewFactory _webViewFactory;
-    private readonly IWebViewService _webViewService;
-    private readonly IWebViewAdapter _webViewAdapter;
-    private readonly IWebViewFocusRegistry _webViewFocusRegistry;
-
-    // Latest edit availability reported by the editor over the bridge. Drives CanPerformEdit.
-    private EditAvailability _editAvailability = EditAvailability.None;
-
     private readonly ContributionDocumentViewModel _viewModel;
-
-    private ContributionDocumentHandler? _documentHandler;
-    private ContributionToolsHandler? _toolsHandler;
-    private IDisposable? _appStateConnection;
-
-    // This view's own state store (writability), mirrored to its WebView over the viewState channel.
-    private IStateStore? _viewState;
-    private IDisposable? _viewStateConnection;
-
-    // JSON-RPC infrastructure. Teardown detaches the WebView2 channel or disposes the deferred
-    // WebSocket channel, depending on the transport the host channel factory selected.
-    private Action? _hostChannelTeardown;
-
-    // Set when the WebSocket transport is in use, to resync the editor after a reconnect.
-    private ProxyHostChannel? _proxyChannel;
-
-    // WebView tool bridge registration tracking. Only set when the package allows the
-    // webview_* tools and the registration has succeeded. The field doubles as a guard
-    // for unregistration.
-    private IDocumentWebViewToolBridge? _toolBridge;
-    private ResourceKey _toolBridgeRegisteredResource;
-
-    // Deferred editor state for views where the WebView initializes asynchronously.
-    // RestoreEditorStateAsync stores state here when the editor isn't ready yet,
-    // and SetContentLoaded applies it once the JS client signals readiness.
-    private string? _pendingEditorStateJson;
-    private bool _isContentLoaded;
-
-    // Save tracking state for async save coordination with WebView
-    private bool _isSaveInProgress;
-    private bool _hasPendingSave;
-
-    // Reload coalescing: at most one external-reload runs at a time. FileSystemWatcher
-    // often emits duplicate Changed events for a single logical write, and the
-    // editor host cannot tolerate concurrent NotifyExternalChangeAsync calls.
-    // Requests arriving while a reload is in flight collapse into a single
-    // follow-up pass.
-    private readonly object _reloadLock = new();
-    private bool _isReloadInProgress;
-    private bool _hasPendingReload;
-
-    // Set when the host channel re-binds a reconnected transport. Forces the next reload to run even
-    // when the disk content matches the file tracking info, because the editor may have missed a reload
-    // notification that was in transit when the previous transport died.
-    private bool _forceReload;
-
-    // Completed by InitContributionViewAsync with the init outcome. LoadContent
-    // triggers the init on first call and awaits this TCS so the open-document
-    // flow returns only when the WebView and host are ready for RPCs.
-    private TaskCompletionSource<Result>? _initTcs;
-
-    /// <summary>
-    /// The WebView2 control acquired from the factory.
-    /// </summary>
-    private WebView2? WebView { get; set; }
-
-    /// <summary>
-    /// The Celbridge host for JSON-RPC communication with the WebView.
-    /// </summary>
-    private CelbridgeHost? Host { get; set; }
+    private readonly ContributionEditorController _controller;
 
     protected override DocumentViewModel DocumentViewModel => _viewModel;
 
@@ -119,32 +27,22 @@ public sealed partial class ContributionDocumentView : DocumentView, IHostInput,
 
     public ContributionDocumentView(
         IServiceProvider serviceProvider,
-        ILogger<ContributionDocumentView> logger,
-        ICommandService commandService,
-        IMessengerService messengerService,
-        IUserInterfaceService userInterfaceService,
-        IStringLocalizer stringLocalizer,
-        IDialogService dialogService,
-        IWebViewFactory webViewFactory,
-        IWebViewService webViewService)
+        IMessengerService messengerService)
     {
-        _logger = logger;
-        _commandService = commandService;
         _messengerService = messengerService;
-        _stringLocalizer = stringLocalizer;
-        _dialogService = dialogService;
-        _userInterfaceService = userInterfaceService;
-        _serviceProvider = serviceProvider;
-        _webViewFactory = webViewFactory;
-        _webViewService = webViewService;
-        _webViewAdapter = ServiceLocator.AcquireService<IWebViewAdapter>();
-        _webViewFocusRegistry = ServiceLocator.AcquireService<IWebViewFocusRegistry>();
-
         _viewModel = serviceProvider.GetRequiredService<ContributionDocumentViewModel>();
 
         this.InitializeComponent();
 
-        _viewModel.ReloadRequested += ViewModel_ReloadRequested;
+        var focusContext = new ContributionEditorFocusContext(
+            WorkspacePanel.Documents,
+            () => _messengerService.Send(new DocumentViewFocusedMessage(_viewModel.FileResource)));
+
+        _controller = new ContributionEditorController(
+            serviceProvider,
+            _viewModel,
+            ContributionWebViewContainer,
+            focusContext);
     }
 
     public override bool HasUnsavedChanges => _viewModel.HasUnsavedChanges;
@@ -156,702 +54,37 @@ public sealed partial class ContributionDocumentView : DocumentView, IHostInput,
 
     protected override async Task<Result> SaveDocumentContentAsync()
     {
-        if (Host is null || _documentHandler is null)
-        {
-            _logger.LogDebug("Save skipped - Host not initialized");
-            return Result.Ok();
-        }
-
-        if (!TryBeginSave())
-        {
-            _logger.LogDebug("Save already in progress, queuing pending save");
-            return Result.Ok();
-        }
-
-        var saveResultTcs = new TaskCompletionSource<Result>();
-        _documentHandler.SaveResultTcs = saveResultTcs;
-
-        await Host.NotifyRequestSaveAsync();
-
-        var timeout = TimeSpan.FromSeconds(SaveRequestTimeoutSeconds);
-        var timeoutTask = Task.Delay(timeout);
-        var completedTask = await Task.WhenAny(saveResultTcs.Task, timeoutTask);
-
-        if (completedTask == timeoutTask)
-        {
-            _documentHandler.SaveResultTcs = null;
-            CompleteSave();
-
-            var errorMessage = $"Contribution editor failed to respond within {SaveRequestTimeoutSeconds} seconds. " +
-                               $"File: {_viewModel.FilePath}";
-
-            _logger.LogError(errorMessage);
-
-            return Result.Fail(errorMessage);
-        }
-
-        var result = await saveResultTcs.Task;
-        _documentHandler.SaveResultTcs = null;
-
-        return result;
-    }
-
-    private bool TryBeginSave()
-    {
-        if (_isSaveInProgress)
-        {
-            _hasPendingSave = true;
-            return false;
-        }
-
-        _isSaveInProgress = true;
-        _hasPendingSave = false;
-        return true;
-    }
-
-    private bool CompleteSave()
-    {
-        _isSaveInProgress = false;
-
-        if (_hasPendingSave)
-        {
-            _hasPendingSave = false;
-            return true;
-        }
-
-        return false;
-    }
-
-    private IContributionEditorLoader ResolveContributionEditorLoader(PackageInfo package)
-    {
-        // The loopback default is registered first and matches every package, so it is the fallback. A
-        // module's custom loader is registered later and wins as the last matching loader.
-        IContributionEditorLoader? selected = null;
-        foreach (var candidate in _serviceProvider.GetServices<IContributionEditorLoader>())
-        {
-            if (candidate.CanLoad(package))
-            {
-                selected = candidate;
-            }
-        }
-
-        Guard.IsNotNull(selected);
-
-        return selected;
-    }
-
-    private async Task InitContributionViewAsync()
-    {
-        if (Contribution is null)
-        {
-            var error = "Cannot initialize contribution view: Contribution is not set";
-            _logger.LogError(error);
-            _initTcs!.TrySetResult(Result.Fail(error));
-            return;
-        }
-
-        _viewModel.Contribution = Contribution;
-
-        var editorLoader = ResolveContributionEditorLoader(Contribution.Package);
-
-        // The factory hands back a control with CoreWebView2 already live (pre-warmed where possible), so
-        // the host is configured immediately and init is signalled once the editor's navigation has started.
-        try
-        {
-            WebView = await _webViewFactory.AcquireAsync();
-            ContributionWebViewContainer.Children.Add(WebView);
-
-            await ConfigureWebViewHostAsync(editorLoader);
-
-            _initTcs!.TrySetResult(Result.Ok());
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, $"Failed to initialize contribution view: {Contribution.Package.Name}");
-            TeardownWebViewState();
-            var failure = Result.Fail($"Failed to initialize contribution view: {Contribution.Package.Name}")
-                .WithException(ex);
-            _initTcs!.TrySetResult(failure);
-        }
-    }
-
-    // Configures a live WebView (CoreWebView2 ready): host channel, RPC targets, tool bridge, navigation gate,
-    // and the editor load. Shared by the pooled and in-place init paths.
-    private async Task ConfigureWebViewHostAsync(IContributionEditorLoader editorLoader)
-    {
-        Guard.IsNotNull(Contribution);
-        Guard.IsNotNull(WebView);
-
-        // DevTools is off when the hosting package blocks it (sensitive material)
-        // or when the user has not enabled the WebViewDevTools feature flag.
-        var devToolsBlocked = Contribution.Package.DevToolsBlocked;
-        WebView.CoreWebView2.Settings.AreDevToolsEnabled =
-            !devToolsBlocked && _webViewService.IsDevToolsFeatureEnabled();
-
-        // Register this editor's web surface; it hosts an edit target (this) for the Edit commands.
-        RegisterWebSurfaceFocus(WebView, editTarget: this, ReleaseFocus);
-
-        // Every contribution editor's assets (its lib, the shared client) are served from the loopback
-        // file server, so register this package's folder there. The resolved loader decides where the
-        // entry page itself loads from.
-        var fileServer = _serviceProvider.GetRequiredService<IFileServer>();
-        var packageUrlName = Contribution.Package.Name.Replace('.', '-');
-
-        fileServer.RegisterPackageFolder(packageUrlName, Contribution.Package.PackageFolder);
-
-        WarnOnEmptyPackageSecrets();
-
-        // Inject the in-page tool bridge shim for the webview_* MCP tool namespace.
-        // Skipped when the package opts out via DevToolsBlocked (sensitive material)
-        // so that no tool surface is exposed for those packages.
-        if (!devToolsBlocked)
-        {
-            await TryInjectToolBridgeShimAsync();
-        }
-
-        // Block all new window requests
-        WebView.CoreWebView2.NewWindowRequested += (s, args) =>
-        {
-            args.Handled = true;
-        };
-
-        WebView.CoreWebView2.ProcessFailed += (s, args) =>
-        {
-            _logger.LogError(
-                "WebView ProcessFailed: Kind={Kind}, Reason={Reason}, ExitCode={ExitCode}",
-                args.ProcessFailedKind, args.Reason, args.ExitCode);
-        };
-
-        // Wire up the JSON-RPC host channel. The loader's declared transport selects between the loopback
-        // WebSocket (the page derives the socket URL from its own origin plus a connection token) and the
-        // WebView2 message channel (for a page that is not same-origin with the loopback server).
-        var useWebSocketChannel = editorLoader.GetTransport(Contribution.Package) == HostChannelTransport.LoopbackWebSocket;
-        var hostChannelBroker = _serviceProvider.GetRequiredService<IHostChannelBroker>();
-        var hostChannelSetup = HostChannelFactory.Create(WebView.CoreWebView2, useWebSocketChannel, hostChannelBroker);
-        _hostChannelTeardown = hostChannelSetup.Teardown;
-        var connectionToken = hostChannelSetup.ConnectionToken;
-        Host = new CelbridgeHost(hostChannelSetup.Channel);
-
-        // A reconnected transport (e.g. after an OS suspend dropped the socket) may have lost messages
-        // that were in transit when the previous socket died, so resync the editor on every rebind.
-        if (hostChannelSetup.Channel is ProxyHostChannel proxyChannel)
-        {
-            _proxyChannel = proxyChannel;
-            proxyChannel.Rebound += OnHostChannelRebound;
-        }
-
-        Host.AddLocalRpcTarget<IHostInput>(this);
-        Host.AddLocalRpcTarget<IHostContext>(this);
-
-        _documentHandler = new ContributionDocumentHandler(
-            _viewModel,
-            _logger,
-            CreateDocumentMetadata,    // Callback to construct document metadata on demand
-            CompleteSave);             // Callback to update state when saving has completed
-
-        _documentHandler.ContentLoaded += SetContentLoaded;
-
-        var dialogHandler = new ContributionDialogHandler(
-            _dialogService,
-            _stringLocalizer,
-            _viewModel);
-
-        Host.AddLocalRpcTarget<IHostDocument>(_documentHandler);
-        Host.AddLocalRpcTarget<IHostDialog>(dialogHandler);
-
-        var mcpToolBridge = _serviceProvider.GetService<IMcpToolBridge>();
-        if (mcpToolBridge is not null)
-        {
-            _toolsHandler = new ContributionToolsHandler(mcpToolBridge, Contribution.Package.PermittedTools);
-            Host.AddLocalRpcTarget<ContributionToolsHandler>(_toolsHandler);
-        }
-
-        Host.StartListening();
-
-        // Registering pushes the current snapshot, so it must run after StartListening. Seed writability
-        // before registering so the connect push carries it.
-        var stateService = _serviceProvider.GetRequiredService<IWebViewStateService>();
-        var capturedHost = Host;
-        _appStateConnection = stateService.AppState.RegisterConnection(
-            snapshot => capturedHost.Rpc.NotifyWithParameterObjectAsync(StateRpcMethods.AppStateChanged, snapshot));
-
-        _viewState = stateService.CreateViewState();
-        _viewState.SetValue("writable", WritableState.ToString());
-        // The preview find bar is built only where the WebView backend has no find bar of its own; where it
-        // does (Chromium's WebView2), the package stays hands-off and Ctrl+F reaches the built-in bar.
-        _viewState.SetValue("providesBuiltInFind", _webViewAdapter.ProvidesBuiltInFind ? "true" : "false");
-        _viewStateConnection = _viewState.RegisterConnection(
-            snapshot => capturedHost.Rpc.NotifyWithParameterObjectAsync(StateRpcMethods.ViewStateChanged, snapshot));
-
-        // Register with the WebView tool bridge so the webview_* MCP tools can
-        // target this WebView by resource key. Mirrors the shim injection guard.
-        if (!devToolsBlocked)
-        {
-            TryRegisterWithToolBridge();
-        }
-
-        var entryPoint = Contribution.EntryPoint;
-        var serverPort = _serviceProvider.GetRequiredService<IServerService>().Port;
-        var loadRequest = new ContributionEditorLoadRequest(
-            WebView!,
-            Contribution.Package,
-            packageUrlName,
-            entryPoint,
-            connectionToken,
-            serverPort);
-
-        // Block all navigations except the editor's own origin. Each allowed navigation also resets the
-        // tool bridge's content-ready gate so webview_* tool calls block until the editor signals
-        // readiness post-navigation.
-        var allowedNavigationPrefix = editorLoader.GetAllowedNavigationOrigin(loadRequest);
-        WebView!.NavigationStarting += (s, args) =>
-        {
-            var uri = args.Uri;
-            if (string.IsNullOrEmpty(uri))
-            {
-                return;
-            }
-
-            if (uri.StartsWith(allowedNavigationPrefix))
-            {
-                _toolBridge?.NotifyContentLoading(_toolBridgeRegisteredResource);
-                return;
-            }
-
-            args.Cancel = true;
-        };
-
-        await editorLoader.LoadAsync(loadRequest);
-    }
-
-    /// <summary>
-    /// Tears down the WebView, host channel, and associated handlers. Safe to call
-    /// multiple times and from partially initialized states.
-    /// </summary>
-    private void TeardownWebViewState()
-    {
-        if (_toolBridge is not null)
-        {
-            _toolBridge.Unregister(_toolBridgeRegisteredResource);
-            _toolBridge = null;
-            _toolBridgeRegisteredResource = ResourceKey.Empty;
-        }
-
-        if (_documentHandler is not null)
-        {
-            _documentHandler.ContentLoaded -= SetContentLoaded;
-        }
-
-        if (WebView is not null)
-        {
-            if (WebView.CoreWebView2 is not null)
-            {
-                _webViewFocusRegistry.Unregister(WebView.CoreWebView2);
-            }
-
-            _webViewAdapter.CloseWebView(WebView, ContributionWebViewContainer);
-
-            WebView = null;
-        }
-
-        if (_proxyChannel is not null)
-        {
-            _proxyChannel.Rebound -= OnHostChannelRebound;
-            _proxyChannel = null;
-        }
-
-        _appStateConnection?.Dispose();
-        _viewStateConnection?.Dispose();
-        Host?.Dispose();
-        _hostChannelTeardown?.Invoke();
-
-        _appStateConnection = null;
-        _viewStateConnection = null;
-        _viewState = null;
-        Host = null;
-        _hostChannelTeardown = null;
-    }
-
-    private async Task TryInjectToolBridgeShimAsync()
-    {
-        var coreWebView2 = WebView?.CoreWebView2;
-        if (coreWebView2 is null)
-        {
-            return;
-        }
-
-        var toolBridge = _serviceProvider.GetService<IDocumentWebViewToolBridge>();
-        if (toolBridge is null)
-        {
-            return;
-        }
-
-        // Install the tool bridge shim as a document-start script so it wraps console/fetch before page
-        // scripts run -- required for get_console / get_network capture. The Skia heads also re-deliver it
-        // per navigation through OnNavigationCompleted_ReinjectShim.
-        try
-        {
-            var script = toolBridge.GetShimScript();
-            await _webViewAdapter.InstallDocumentStartScriptAsync(coreWebView2, script);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(ex, "Failed to install the document-start WebView tool bridge shim");
-        }
-
-        coreWebView2.NavigationCompleted += OnNavigationCompleted_ReinjectShim;
-    }
-
-    private async void OnNavigationCompleted_ReinjectShim(CoreWebView2 sender, CoreWebView2NavigationCompletedEventArgs args)
-    {
-        // async void event handler: swallow exceptions to protect the process. The bridge caches the shim
-        // after the first read, and the shim is idempotent. Re-injection is a no-op on Windows, where the
-        // document-start script persists across navigations.
-        if (_toolBridge is null)
-        {
-            return;
-        }
-
-        try
-        {
-            var script = _toolBridge.GetShimScript();
-            await _webViewAdapter.ReinjectDocumentStartScriptAsync(sender, script);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(ex, "Failed to re-inject the WebView tool bridge shim");
-        }
-    }
-
-    private void TryRegisterWithToolBridge()
-    {
-        var webView = WebView;
-        if (webView?.CoreWebView2 is null)
-        {
-            return;
-        }
-
-        var toolBridge = _serviceProvider.GetService<IDocumentWebViewToolBridge>();
-        if (toolBridge is null)
-        {
-            return;
-        }
-
-        var resource = FileResource;
-        if (resource.IsEmpty)
-        {
-            return;
-        }
-
-        toolBridge.RegisterWebView2(resource, webView, _webViewAdapter);
-
-        _toolBridge = toolBridge;
-        _toolBridgeRegisteredResource = resource;
-    }
-
-    public void OnKeyboardShortcut(string key, bool ctrlKey, bool shiftKey, bool altKey)
-    {
-        var keyboardShortcutService = ServiceLocator.AcquireService<IKeyboardShortcutService>();
-        keyboardShortcutService.HandleShortcut(key, ctrlKey, shiftKey, altKey);
+        return await _controller.SaveContentAsync();
     }
 
     protected override void OnWritableStateChanged()
     {
-        // The store may not exist yet (SetWritableState runs before LoadContent). The seed at registration
-        // captures the current value, so an early change before connect is not lost.
-        _viewState?.SetValue("writable", WritableState.ToString());
-    }
-
-    private DocumentMetadata CreateDocumentMetadata()
-    {
-        var locale = System.Globalization.CultureInfo.CurrentUICulture.TwoLetterISOLanguageName;
-
-        var metaData = new DocumentMetadata(
-            DocumentViewModel.FilePath,
-            DocumentViewModel.FileResource.ToString(),
-            Path.GetFileName(DocumentViewModel.FilePath),
-            locale);
-
-        return metaData;
-    }
-
-    private async void SetContentLoaded(ContentLoadedReason reason = ContentLoadedReason.Initial)
-    {
-        // Tool bridge readiness fires for every reason — initial load, external reload,
-        // and programmatic webview_reload — so gated webview_* calls unblock as soon
-        // as the editor signals it has reinitialised post-navigation.
-        _toolBridge?.NotifyContentReady(_toolBridgeRegisteredResource);
-
-        if (reason != ContentLoadedReason.Initial)
-        {
-            return;
-        }
-
-        _isContentLoaded = true;
-
-        if (_pendingEditorStateJson is null)
-        {
-            return;
-        }
-
-        var state = _pendingEditorStateJson;
-        _pendingEditorStateJson = null;
-
-        try
-        {
-            await RestoreEditorStateAsync(state);
-        }
-        catch (Exception ex)
-        {
-            // Editor state restoration is best-effort: a corrupt or incompatible state should
-            // never tear down the process. Log and swallow to preserve the async void safety contract.
-            _logger.LogError(ex, "Failed to restore editor state after content loaded");
-        }
-    }
-
-    private async Task ReloadWithStatePreservationAsync()
-    {
-        if (Host is null || _documentHandler is null)
-        {
-            return;
-        }
-
-        // Resolve the workspace-scoped documents service at call time, then drain
-        // any reload hint registered by the command that triggered this reload.
-        var workspaceWrapper = _serviceProvider.GetRequiredService<IWorkspaceWrapper>();
-        var documentsService = workspaceWrapper.WorkspaceService.DocumentsService;
-        var hint = documentsService.ConsumeReloadHint(_viewModel.FileResource);
-        bool preserveViewState = hint == ReloadHint.PreserveViewState;
-
-        string? savedState = null;
-        try
-        {
-            savedState = await TrySaveEditorStateAsync();
-        }
-        catch (Exception ex)
-        {
-            _logger.LogDebug(ex, "Failed to capture editor state before external reload; continuing without preservation");
-        }
-
-        var reloadComplete = new TaskCompletionSource();
-        void OnReloaded(ContentLoadedReason reason)
-        {
-            if (reason == ContentLoadedReason.ExternalReload)
-            {
-                reloadComplete.TrySetResult();
-            }
-        }
-        _documentHandler.ContentLoaded += OnReloaded;
-
-        try
-        {
-            await Host.NotifyExternalChangeAsync(preserveViewState);
-
-            var completed = await Task.WhenAny(reloadComplete.Task, Task.Delay(TimeSpan.FromSeconds(ReloadStateWaitSeconds)));
-            if (completed != reloadComplete.Task)
-            {
-                // Expected while the editor transport is mid-reconnect; the rebind resync re-runs the reload.
-                _logger.LogWarning(
-                    "Editor did not confirm external reload within {Seconds}s. File: {File}",
-                    ReloadStateWaitSeconds, _viewModel.FilePath);
-                return;
-            }
-
-            if (!string.IsNullOrEmpty(savedState))
-            {
-                await RestoreEditorStateAsync(savedState);
-            }
-        }
-        finally
-        {
-            _documentHandler.ContentLoaded -= OnReloaded;
-        }
-    }
-
-    public override async Task<string?> TrySaveEditorStateAsync()
-    {
-        if (Host is null || !_isContentLoaded)
-        {
-            return null;
-        }
-
-        // Race the request against a hard timeout and abandon it on timeout. A CancellationToken does
-        // not work here: StreamJsonRpc cancellation waits for the editor to acknowledge the cancel, and
-        // the unresponsive editor is exactly the failure being guarded against. State capture is
-        // best-effort, so abandoning the request and closing without state is the correct degradation.
-        var requestStateTask = Host.RequestStateAsync();
-        var timeoutTask = Task.Delay(TimeSpan.FromSeconds(EditorStateRequestTimeoutSeconds));
-        var completedTask = await Task.WhenAny(requestStateTask, timeoutTask);
-
-        if (completedTask != requestStateTask)
-        {
-            _logger.LogWarning("Editor did not return state within {Seconds}s; closing without preserving editor state.", EditorStateRequestTimeoutSeconds);
-            ObserveAbandonedRequest(requestStateTask);
-            return null;
-        }
-
-        try
-        {
-            return await requestStateTask;
-        }
-        catch (StreamJsonRpc.RemoteMethodNotFoundException)
-        {
-            // Editor did not register a document/requestState handler.
-            return null;
-        }
-    }
-
-    // Swallows the eventual fault of an abandoned host->editor request (it faults when the WebView is
-    // torn down) so it does not surface as an unobserved task exception.
-    private static void ObserveAbandonedRequest(Task task)
-    {
-        _ = task.ContinueWith(
-            static abandonedTask => { _ = abandonedTask.Exception; },
-            CancellationToken.None,
-            TaskContinuationOptions.OnlyOnFaulted,
-            TaskScheduler.Default);
-    }
-
-    public override async Task RestoreEditorStateAsync(string state)
-    {
-        if (!_isContentLoaded)
-        {
-            _pendingEditorStateJson = state;
-            return;
-        }
-
-        var restoreStateTask = Host!.RestoreStateAsync(state);
-        var timeoutTask = Task.Delay(TimeSpan.FromSeconds(EditorStateRequestTimeoutSeconds));
-        var completedTask = await Task.WhenAny(restoreStateTask, timeoutTask);
-
-        if (completedTask != restoreStateTask)
-        {
-            // Best-effort restore. An unresponsive editor should not stall the caller. Abandon and move on.
-            _logger.LogWarning("Editor did not acknowledge restoreState within {Seconds}s; continuing.", EditorStateRequestTimeoutSeconds);
-            ObserveAbandonedRequest(restoreStateTask);
-            return;
-        }
-
-        try
-        {
-            await restoreStateTask;
-        }
-        catch (StreamJsonRpc.RemoteMethodNotFoundException)
-        {
-            // Editor did not register a document/restoreState handler.
-        }
-    }
-
-    public void OnLinkClicked(string href)
-    {
-        if (string.IsNullOrEmpty(href))
-        {
-            return;
-        }
-
-        if (Contribution is null)
-        {
-            return;
-        }
-
-        var resolveResult = _viewModel.ResolveLinkTarget(href);
-
-        if (resolveResult.IsFailure)
-        {
-            _logger.LogWarning($"Failed to resolve link: {href}");
-            _ = ShowLinkErrorAsync(href);
-            return;
-        }
-
-        var resourceKey = resolveResult.Value;
-
-        if (resourceKey.IsEmpty)
-        {
-            OpenSystemBrowser(_commandService, href);
-        }
-        else
-        {
-            _commandService.Execute<IOpenDocumentCommand>(command =>
-            {
-                command.FileResource = resourceKey;
-            });
-        }
-    }
-
-    private static void OpenSystemBrowser(ICommandService commandService, string? uri)
-    {
-        if (string.IsNullOrEmpty(uri))
-        {
-            return;
-        }
-
-        commandService.Execute<IOpenBrowserCommand>(command =>
-        {
-            command.URL = uri;
-        });
-    }
-
-    private async Task ShowLinkErrorAsync(string href)
-    {
-        var errorTitle = _stringLocalizer.GetString("Extension_LinkError_Title");
-        var errorMessage = _stringLocalizer.GetString("Extension_LinkError_Message", href);
-        await _dialogService.ShowAlertDialogAsync(errorTitle, errorMessage);
+        _controller.SetWritableState(WritableState);
     }
 
     public override async Task<Result> LoadContent()
     {
-        // LoadContent drives WebView initialization directly — not via the
-        // XAML Loaded event. DocumentsService.CreateDocumentView calls
-        // LoadContent before the view is added to the visual tree, so a
-        // Loaded-driven init would not fire in time to navigate the WebView.
-        // We wait for init (host ready, navigation started) but do not block
-        // on the editor's own notifyContentLoaded signal. Some heavyweight
-        // editors can take seconds to finish importing content. Forcing callers
-        // to wait for that would make document open feel slow.
-        if (_initTcs is null)
+        if (Contribution is null)
         {
-            _initTcs = new TaskCompletionSource<Result>();
-            _ = InitContributionViewAsync();
+            return Result.Fail("Cannot initialize contribution view: Contribution is not set");
         }
 
-        var initResult = await _initTcs.Task;
-        if (initResult.IsFailure)
-        {
-            return initResult;
-        }
-
-        return Result.Ok();
+        return await _controller.InitializeAsync(Contribution);
     }
 
     public override async Task<Result> NavigateToLocation(string location)
     {
-        if (Host is null ||
-            string.IsNullOrEmpty(location))
-        {
-            return Result.Ok();
-        }
+        return await _controller.NavigateToLocationAsync(location);
+    }
 
-        try
-        {
-            using var doc = JsonDocument.Parse(location);
-            var root = doc.RootElement;
+    public override Task<string?> TrySaveEditorStateAsync()
+    {
+        return _controller.TrySaveEditorStateAsync();
+    }
 
-            var lineNumber = root.TryGetProperty("lineNumber", out var lineProp) ? lineProp.GetInt32() : 1;
-            var column = root.TryGetProperty("column", out var colProp) ? colProp.GetInt32() : 1;
-            var endLineNumber = root.TryGetProperty("endLineNumber", out var endLineProp) ? endLineProp.GetInt32() : 0;
-            var endColumn = root.TryGetProperty("endColumn", out var endColProp) ? endColProp.GetInt32() : 0;
-
-            await Host.Rpc.NotifyWithParameterObjectAsync(
-                "editor/navigateToLocation",
-                new { lineNumber, column, endLineNumber, endColumn });
-
-            return Result.Ok();
-        }
-        catch (Exception ex)
-        {
-            return Result.Fail($"Failed to navigate to location: {location}")
-                .WithException(ex);
-        }
+    public override Task RestoreEditorStateAsync(string state)
+    {
+        return _controller.RestoreEditorStateAsync(state);
     }
 
     public override async Task<bool> CanClose()
@@ -865,288 +98,17 @@ public sealed partial class ContributionDocumentView : DocumentView, IHostInput,
 
     public override async Task PrepareToClose()
     {
-        _isContentLoaded = false;
-        _pendingEditorStateJson = null;
-
         _messengerService.UnregisterAll(this);
 
-        _viewModel.ReloadRequested -= ViewModel_ReloadRequested;
+        _controller.Teardown();
 
         _viewModel.Cleanup();
-
-        TeardownWebViewState();
 
         await base.PrepareToClose();
     }
 
     public override void FocusDocument()
     {
-        // A tab click focuses the web content (native first responder on macOS, where no managed GotFocus
-        // follows). The registry gives it focus and reports it, releasing the previously focused surface.
-        if (WebView is not null)
-        {
-            _webViewFocusRegistry.GrantFocus(WebView);
-        }
-    }
-
-    private void ReleaseFocus()
-    {
-        _ = Host?.NotifyReleaseFocusAsync();
-    }
-
-    public bool CanPerformEdit(EditIntent intent)
-    {
-        return _editAvailability.Allows(intent);
-    }
-
-    public void PerformEdit(EditIntent intent)
-    {
-        // The WebView's own JS clipboard write is blocked outside a user gesture on the Skia WKWebView,
-        // so the clipboard verbs are host-mediated: the host moves text between Monaco and the native
-        // clipboard. Select-all, undo, and redo touch no clipboard and run inside the editor.
-        switch (intent)
-        {
-            case EditIntent.Copy:
-                _ = CopyEditorSelectionAsync(deleteSelection: false);
-                break;
-
-            case EditIntent.Cut:
-                _ = CopyEditorSelectionAsync(deleteSelection: true);
-                break;
-
-            case EditIntent.Paste:
-                _ = PasteIntoEditorAsync();
-                break;
-
-            case EditIntent.SelectAll:
-                _ = Host?.NotifyPerformEditAsync("selectAll");
-                break;
-
-            case EditIntent.Undo:
-                _ = Host?.NotifyPerformEditAsync("undo");
-                break;
-
-            case EditIntent.Redo:
-                _ = Host?.NotifyPerformEditAsync("redo");
-                break;
-        }
-    }
-
-    public bool TryHandleTabKey(bool shift)
-    {
-        // A code editor with text focus indents or outdents. It reports that over the bridge, so a read-only
-        // or unfocused editor falls through to the generic Tab notification below.
-        if (_editAvailability.CanIndent)
-        {
-            var command = shift ? "outdent" : "indent";
-            _ = Host?.NotifyPerformEditAsync(command);
-
-            return true;
-        }
-
-        // Other editors handle Tab their own way (the spreadsheet moves the active cell). Editors that do
-        // not act on it ignore the notification, and the key is still swallowed so focus stays in the document.
-        _ = Host?.NotifyTabKeyAsync(shift);
-
-        return true;
-    }
-
-    private async Task CopyEditorSelectionAsync(bool deleteSelection)
-    {
-        var host = Host;
-        if (host is null)
-        {
-            return;
-        }
-
-        try
-        {
-            var selectedText = await host.Rpc.InvokeAsync<string?>("editor/getSelectedText");
-            if (string.IsNullOrEmpty(selectedText))
-            {
-                return;
-            }
-
-            _commandService.Execute<ICopyTextToClipboardCommand>(command => command.Text = selectedText);
-
-            if (deleteSelection)
-            {
-                await host.Rpc.NotifyWithParameterObjectAsync("editor/insertText", new { text = string.Empty });
-            }
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Failed to copy the editor selection to the clipboard");
-        }
-    }
-
-    private async Task PasteIntoEditorAsync()
-    {
-        var host = Host;
-        if (host is null)
-        {
-            return;
-        }
-
-        try
-        {
-            var dataPackageView = Clipboard.GetContent();
-            if (!dataPackageView.Contains(StandardDataFormats.Text))
-            {
-                return;
-            }
-
-            var text = await dataPackageView.GetTextAsync();
-            await host.Rpc.NotifyWithParameterObjectAsync("editor/insertText", new { text });
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Failed to paste the clipboard into the editor");
-        }
-    }
-
-    public void OnEditAvailabilityChanged(
-        bool canCopy,
-        bool canCut,
-        bool canPaste,
-        bool canSelectAll,
-        bool canUndo,
-        bool canRedo,
-        bool canIndent = false)
-    {
-        _editAvailability = new EditAvailability(
-            canCopy,
-            canCut,
-            canPaste,
-            canSelectAll,
-            canUndo,
-            canRedo,
-            canIndent);
-    }
-
-    private void OnHostChannelRebound(object? sender, EventArgs e)
-    {
-        lock (_reloadLock)
-        {
-            _forceReload = true;
-        }
-
-        // Raised on the WebSocket endpoint's request thread; marshal to the UI thread where the
-        // reload pipeline runs.
-        DispatcherQueue.TryEnqueue(() => ViewModel_ReloadRequested(this, EventArgs.Empty));
-    }
-
-    private async void ViewModel_ReloadRequested(object? sender, EventArgs e)
-    {
-        // Coalesce concurrent reload requests. FileSystemWatcher commonly emits
-        // duplicate Changed events for one logical write. A second reload arriving
-        // mid-flight folds into one follow-up pass instead of racing the first.
-        lock (_reloadLock)
-        {
-            if (_isReloadInProgress)
-            {
-                _hasPendingReload = true;
-                return;
-            }
-            _isReloadInProgress = true;
-        }
-
-        while (true)
-        {
-            lock (_reloadLock)
-            {
-                _forceReload = false;
-            }
-
-            // async void: catch everything so a faulty editor cannot crash the process.
-            try
-            {
-                await ReloadWithStatePreservationAsync();
-                // Sync the ViewModel's external-change tracking with the disk
-                // content we just loaded so duplicate watcher events for this
-                // write hash-match the cache on the next iteration.
-                await _viewModel.UpdateFileTrackingInfoAsync();
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "External reload failed for contribution document");
-            }
-
-            // Drain any pending request. Skip the follow-up reload when the disk
-            // content has not actually changed since the reload we just ran (the
-            // duplicate-watcher-event case) -- unless a transport rebind forced it,
-            // in which case the editor may be stale even though the tracking info
-            // matches the disk.
-            bool runFollowUp = false;
-            while (!runFollowUp)
-            {
-                bool wasPending;
-                bool forceReload;
-                lock (_reloadLock)
-                {
-                    wasPending = _hasPendingReload;
-                    _hasPendingReload = false;
-                    forceReload = _forceReload;
-                    if (!wasPending)
-                    {
-                        _isReloadInProgress = false;
-                        return;
-                    }
-                }
-
-                if (forceReload)
-                {
-                    runFollowUp = true;
-                    continue;
-                }
-
-                try
-                {
-                    runFollowUp = await _viewModel.IsFileChangedExternallyAsync();
-                }
-                catch (Exception ex)
-                {
-                    // Treat a failed disk probe as "assume changed" so we don't
-                    // silently drop a legitimately-queued reload.
-                    _logger.LogDebug(ex, "External change probe failed; running follow-up reload defensively");
-                    runFollowUp = true;
-                }
-            }
-        }
-    }
-
-    private void WarnOnEmptyPackageSecrets()
-    {
-        Guard.IsNotNull(Contribution);
-
-        // An empty secret value almost certainly indicates a missing private license file or a module that
-        // failed to populate its BundledPackageDescriptor. The editor at the other end will typically fail
-        // to activate, so surface it loudly here.
-        foreach (var pair in Contribution.Package.Secrets)
-        {
-            if (string.IsNullOrEmpty(pair.Value))
-            {
-                _logger.LogWarning(
-                    "Secret '{SecretName}' for package '{PackageName}' is empty; the editor will likely fail to activate.",
-                    pair.Key, Contribution.Package.Name);
-            }
-        }
-    }
-
-    // Builds the capability context (permitted tools, secrets, options) that the JS client fetches over the
-    // bridge via host/getContext on every head.
-    public CelbridgeContext GetContext()
-    {
-        return BuildCelbridgeContext();
-    }
-
-    private CelbridgeContext BuildCelbridgeContext()
-    {
-        Guard.IsNotNull(Contribution);
-
-        return new CelbridgeContext(
-            Contribution.Package.PermittedTools,
-            Contribution.Package.Secrets,
-            Contribution.Options);
+        _controller.FocusWebView();
     }
 }
